+++ frontend/app/src/components/history/transactions/TransactionContent.vue
@@ -3,7 +3,6 @@ import { type Account, type GeneralAccount } from '@rotki/common/lib/account';
 import { type ComputedRef, type Ref } from 'vue';
 import { type DataTableHeader } from 'vuetify';
 import { type BlockchainSelection } from '@rotki/common/lib/blockchain';
-import { omit } from 'lodash';
 import {
   type HistoryEventType,
   type TransactionEventProtocol
@@ -234,7 +233,7 @@ const saveData = async (
   event: NewEthTransactionEvent | EthTransactionEventEntry
 ) => {
   if ('identifier' in event) {
-    return await editTransactionEvent(omit(event, 'evmChain'));
+    return await editTransactionEvent(event);
   }
   return await addTransactionEvent(event);
 };